Sentiment-ChatBot/requirements.txt
google-generativeai>=0.3.0 
streamlit>=1.10.0 
python-dotenv>=0.19.0
pandas>=1.3.0